Fix GCC build
git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/trunk@16206 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859

+++ src/jrd/vio.cpp
@@ -5065,7 +5065,8 @@ void RelationGarbage::getGarbage(const SLONG oldest_snapshot, PageBitmap **sbm)
 		}
 		delete garbage.bm;
 
-		array.remove(0U);
+		// Need to cast zero to exact type because literal zero means null pointer
+		array.remove(static_cast<size_t>(0));
 	}
 }
